refactor(suite): useAccounts to selector
