+++ packages/suite/src/hooks/wallet/index.ts
@@ -1,2 +1,1 @@
-export { useAccounts } from './useAccounts';
 export { useSendFormContext } from './useSendForm';
