+++ packages/suite/src/views/dashboard/AssetsView/AssetsView.tsx
@@ -10,6 +10,7 @@ import {
     isNetworkSymbol,
 } from '@suite-common/wallet-config';
 import {
+    selectAllAccountsToList,
     selectCurrentFiatRates,
     selectEnabledNetworks,
     selectLocalCurrency,
@@ -33,7 +34,6 @@ import { DashboardSection } from 'src/components/dashboard';
 import { Translation } from 'src/components/suite';
 import { useNetworkSupport } from 'src/hooks/settings/useNetworkSupport';
 import { useDiscovery, useDispatch, useLayoutSize, useSelector } from 'src/hooks/suite';
-import { useAccounts } from 'src/hooks/wallet';
 import { Account } from 'src/types/wallet';
 import { selectDiscoveryOverallStatus } from 'src/utils/wallet/selectDiscoveryOverallStatus';
 
@@ -93,7 +93,7 @@ export const AssetsView = () => {
     const dispatch = useDispatch();
     const { isDiscoveryRunning } = useDiscovery();
     const discoveryStatus = useSelector(selectDiscoveryOverallStatus);
-    const accounts = useAccounts();
+    const accounts = useSelector(selectAllAccountsToList);
     const { supportedMainnets } = useNetworkSupport();
     const { isBelowTablet } = useLayoutSize();
